[IMP] connector_office_hr_leave
When a manager confirms a leave, sometimes the token has expired and
this can generate an error.

This PR changes the logic to check the expiration date of the token
before synchronizing to Office365 and in case the token is expired we
request a login. When the token is refreshed, the pending
synchronizations are performed.